extends simple logger
