+++ src/plugins/simple-logger/base.ts
@@ -1,0 +1,9 @@
+
+export enum SimpleLogLevel {
+  TRACE = 0,
+  DEBUG = 1,
+  INFO = 2,
+  WARN = 3,
+  ERROR = 4,
+  FATAL = 5
+}
